basic archive/stream architecture

+++ source/decoders/src/dc6.cpp
@@ -4,8 +4,8 @@
 
 #include "dc6.h"
 #include <fmt/format.h>
-#include <utils.h>
 #include "palette.h"
+#include "utils.h"
 
 DC6::~DC6()
 {
@@ -22,18 +22,19 @@ void DC6::Decode(const char* filename)
 
 bool DC6::extractHeaders()
 {
-    if(fread(&header, sizeof(header), 1, file) != 1)return false;
-    size_t framesNumber = static_cast<size_t>(header.directions * header.frames_per_dir);
+    if (fread(&header, sizeof(header), 1, file) != 1) return false;
+    size_t framesNumber = header.directions * header.frames_per_dir;
     frameHeaders.resize(framesNumber);
 
     framePointers.resize(framesNumber);
-    if(fread(framePointers.data(), sizeof(uint32_t), framesNumber, file) != framesNumber)return false;
+    if (fread(framePointers.data(), sizeof(uint32_t), framesNumber, file) != framesNumber)
+        return false;
 
     for (size_t i = 0; i < framesNumber; ++i)
     {
         FrameHeader& frameHeader = frameHeaders[i];
         fseek(file, framePointers[i], SEEK_SET);
-        if(fread(&frameHeader, sizeof(frameHeader), 1, file) != 1)return false;
+        if (fread(&frameHeader, sizeof(frameHeader), 1, file) != 1) return false;
         fmt::print("\nframe {}\n", i);
         fmt::print("flip {}\n", frameHeader.flip);
         fmt::print("width {}\n", frameHeader.width);
